back_test/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.6.0
python-dotenv==1.0.0
requests>=2.31.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0,<2.0.0
google-generativeai>=0.3.0
openai>=1.0.0
streamlit>=1.28.0